+++ engine/execution/ingestion/blocks/queue_test.go
@@ -1,4 +1,4 @@
-package block_queue
+package blocks
 
 import (
 	"errors"
